+++ Sources/Classes/core/Downloader/MiniAppDownloader.swift
@@ -52,7 +52,7 @@ class MiniAppDownloader: MiniAppDownloaderInterface {
         }
     }
 
-    private func download(appId: String, versionId: String, completionHandler: @escaping (Result<URL, MASDKError>) -> Void) {
+    internal func download(appId: String, versionId: String, completionHandler: @escaping (Result<URL, MASDKError>) -> Void) {
         let miniAppStoragePath = FileManager.getMiniAppVersionDirectory(with: appId, and: versionId)
         self.manifestDownloader.fetchManifest(apiClient: self.miniAppClient, appId: appId, versionId: versionId) { (result) in
             switch result {
@@ -73,6 +73,28 @@ class MiniAppDownloader: MiniAppDownloaderInterface {
                 if error.isQPSLimitError() {
                     MiniAppStorage.cleanVersions(for: appId)
                 }
+                completionHandler(.failure(error))
+            }
+        }
+    }
+
+    internal func downloadMiniApp(appId: String, versionId: String, completionHandler: @escaping (Result<URL, MASDKError>) -> Void) {
+        let miniAppStoragePath = FileManager.getMiniAppVersionDirectory(with: appId, and: versionId)
+        self.manifestDownloader.fetchManifest(apiClient: self.miniAppClient, appId: appId, versionId: versionId) { (result) in
+            switch result {
+            case .success(let responseData):
+                self.startDownloadingFiles(urls: responseData.manifest, to: miniAppStoragePath, miniAppId: appId, miniAppVersion: versionId) { downloadResult in
+                    switch downloadResult {
+                    case .success:
+                        DispatchQueue.main.async {
+                            self.cacheVerifier.storeHash(for: appId, version: versionId)
+                        }
+                        fallthrough
+                    default:
+                        completionHandler(downloadResult)
+                    }
+                }
+            case .failure(let error):
                 completionHandler(.failure(error))
             }
         }