+++ Sources/Classes/core/MiniApp.swift
@@ -206,8 +206,33 @@ public class MiniApp: NSObject {
         try? MiniAppSecureStorage.wipeSecureStorage(for: miniAppId)
     }
 
-    public static func unzipMiniApp(fileName: String, miniAppId: String, versionId: String) {
-        MiniAppSDKUtility.unzipMiniApp(fileName: fileName, miniAppId: miniAppId, versionId: versionId)
+    /// Unzip the provided archive file from bundle.
+    /// - Parameters:
+    ///   - fileName: File name of the Bundle that you want to extract in MiniApp folder
+    ///   - miniAppId: MiniApp ID
+    ///   - versionId: Version ID
+    ///   - completionHandler: Status of the Unzip feature
+    public static func unzipMiniApp(fileName: String, miniAppId: String, versionId: String, completionHandler: ((Result<Bool, MASDKError>) -> Void)? = nil) {
+        MiniAppSDKUtility.unzipMiniApp(fileName: fileName, miniAppId: miniAppId, versionId: versionId, completionHandler: completionHandler)
+    }
+
+    /// Download a specific MiniApp from the Platform, please note that the MiniApp should be available before start downloading
+    /// - Parameters:
+    ///   - appId: MiniApp ID
+    ///   - versionId: VersionID of the MiniApp
+    ///   - completionHandler: Completion handler that tells the status of the download
+    public func downloadMiniApp(appId: String, versionId: String, completionHandler: @escaping (Result<Bool, MASDKError>) -> Void) {
+        realMiniApp.downloadMiniApp(appId: appId, versionId: versionId, completionHandler: completionHandler)
+    }
+
+    /// Check and return TRUE if the MiniApp is available for a given MiniAppID and VersionID.
+    /// This method not only checks if the folder is available, but also checks if index.html is available.
+    /// - Parameters:
+    ///   - appId: MiniApp ID
+    ///   - versionId: VersionID of the MiniApp
+    /// - Returns: TRUE if MiniApp is available
+    public static func isMiniAppCacheAvailable(appId: String, versionId: String) -> Bool {
+        return MiniAppSDKUtility.isMiniAppAvailable(appId: appId, versionId: versionId)
     }
 }